fourth

+++ ft_strcmp.c
@@ -20,9 +20,9 @@ int	ft_strcmp(const char *s1, const char *s2)
 	while (*ch1 != '\0' && *ch2 != '\0')
 	{
 		if (*ch1 != *ch2)
-			return (*ch1 - *ch2);
+			return ((int)(*ch1 - *ch2));
 		ch1++;
 		ch2++;
 	}
-	return (*ch1 - *ch2);
+	return ((int)(*ch1 - *ch2));
 }
